Assignment 6
Run either linear regression on the housing data, or logistic regression with the moons data, with gradient descent with two options: manually computing the gradients and using autodiff.

Use tf.get_default_graph() and tf.Graph.get_operations() to see the similarity and difference between the graphs generated with and without autodiff.  How many nodes are there in these graphs?  Draw these graphs with all the nodes and edges.

Submit a screenshot of your construction and execution of the graphs and also the drawings of the graphs.